+++ src/features/modal/login/LoginModal.stories.tsx
@@ -1,11 +1,11 @@
 import { Meta, StoryObj } from '@storybook/react';
 import { MemoryRouter } from 'react-router-dom';
-import StudyCreationModal from '@/features/studyCreation/StudyCreationModal';
+import LoginModal from '@/features/modal/login/LoginModal';
 import { useState } from 'react';
 
-const meta: Meta<typeof StudyCreationModal> = {
-  title: 'Features/studyCreation/StudyCreationModal',
-  component: StudyCreationModal,
+const meta: Meta<typeof LoginModal> = {
+  title: 'Features/Modal/Login/LoginModal',
+  component: LoginModal,
   decorators: [
     (Story) => (
       <MemoryRouter>
@@ -25,13 +25,13 @@ const meta: Meta<typeof StudyCreationModal> = {
 
 export default meta;
 
-type Story = StoryObj<typeof StudyCreationModal>;
+type Story = StoryObj<typeof LoginModal>;
 
 export const Default: Story = {
   render: (args) => {
     const [open, setOpen] = useState(args.open);
     return (
-      open ? <StudyCreationModal open={open} onClose={() => setOpen(false)} />
+      open ? <LoginModal open={open} onClose={() => setOpen(false)} />
       : <button onClick={() => setOpen(true)}>Open</button>
     );
   },
