+++ src/features/modal/studyCreation/LeftSection.tsx
