+++ src/features/modal/studyCreation/RightSection.tsx
