+++ src/features/modal/studyCreation/StudyCreation.styles.ts
